Make new definitions discoverable
